adding more time stamps at all sorts of places

+++ include/poledata.h
@@ -5,11 +5,11 @@
 typedef struct
 {
   // position of the upper blob in m
-  double pos_upper[DIM];
+  float pos_upper[DIM];
   // position of the lower blob in m
-  double pos_lower[DIM];
+  float pos_lower[DIM];
   // anlew of the pole in rad
-  double pole_angle;
+  float pole_angle;
   // angular velocity of pole [rad/s]
-  double pole_angular_velocity;
+  float pole_angular_velocity;
 } PoleData;
